move options in BesuCommand
